+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update kong setup to make it work with Gateway API.
+
 ## [0.2.1] - 2026-04-16
 
 ### Fixed