+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-04-30
+
+### Changed
+
+- Update kong setup to make it work with Gateway API.
+- Add kong app and user values configmap to kustomization.
+- Update `envoy-loadtesting` so that kong and nginx are not deployed at the same time on the workload cluster. Only one of those is selected by the user.
+
 ## [0.2.1] - 2026-04-16
 
 ### Fixed
@@ -28,7 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial version of the app.
 
-[Unreleased]: https://github.com/giantswarm/microservices-demo-app/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/giantswarm/microservices-demo-app/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/giantswarm/microservices-demo-app/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/giantswarm/microservices-demo-app/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/giantswarm/microservices-demo-app/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/giantswarm/microservices-demo-app/releases/tag/v0.1.0